# docs/openapi.yaml
openapi: 3.0.0
info:
  title: MyTrello API
  version: 0.1.9
  contact:
    name: Shagalov Veaceslav
    url: t.me/slava_shagalov
servers:
  - url: http://localhost/api/v1
    description: Main server
...
paths:
  /auth/signup:
    post:
      summary: Creates new user and returns authentication cookie.
      description: Creates new user and returns authentication cookie.
      security: [ ]
      tags:
        - Auth
      requestBody:
        required: true
        description: Sign up params.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignUpReq"
      responses:
        "200":
          description: >
            Successfully created user.
            The session ID is returned in a cookie named `JSESSIONID`. 
            You need to include this cookie in subsequent requests.
            A JSON with user data in body.
          headers:
            Set-Cookie:
              schema:
                type: string
                example: JSESSIONID=abcde12345; Path=/; HttpOnly
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "500":
          $ref: "#/components/responses/ErrInternalServerError"
  /auth/signin:
    post:
      summary: Logs in and returns the authentication cookie
      description: Logs in and returns the authentication cookie
      requestBody:
        required: true
        description: Sign in params.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignInReq"
      security: [ ]
      tags:
        - Auth
      responses:
        "200":
          description: >
            Successfully authenticated.
            The session ID is returned in a cookie named `JSESSIONID`. 
            You need to include this cookie in subsequent requests.
            A JSON with user data in body.
          headers:
            Set-Cookie:
              schema:
                type: string
                example: JSESSIONID=abcde12345; Path=/; HttpOnly
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "404":
          $ref: "#/components/responses/ErrUserNotFound"
        "500":
          $ref: "#/components/responses/ErrInternalServerError"
  /auth/logout:
    delete:
      summary: Logs out and deletes the authentication cookie.
      description: Logs out and deletes the authentication cookie.
      security:
        - cookieAuth: [ ]
      tags:
        - Auth
      responses:
        "204":
          description: Successfully logged out.
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "500":
          $ref: "#/components/responses/ErrInternalServerError"
...
  /workspaces/{id}/boards:
    parameters:
      - schema:
          type: integer
        name: id
        in: path
        required: true
    post:
      tags:
        - Workspaces
      summary: Create a new board
      description: Create a new board
      requestBody:
        description: A JSON object containing the data to create board.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBoard"
      responses:
        "200":
          description: A JSON object of created board.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Board"
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "500":
          $ref: "#/components/responses/ErrInternalServerError"
      security:
        - cookieAuth: [ ]
    get:
      tags:
        - Workspaces
      summary: Returns boards by workspace id
      description: Returns boards by workspace id
      responses:
        "200":
          description: A JSON with boards array
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Boards"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "500":
          $ref: "#/components/responses/ErrInternalServerError"
      security:
        - cookieAuth: [ ]
  /boards:
    parameters:
      - name: title
        in: query
        description: Query string for board title
        required: false
        schema:
          type: string
    get:
      tags:
        - Boards
      summary: Returns user boards by filter parameters
      description: Returns user boards by filter parameters
      responses:
        "200":
          description: A JSON object with filtered user boards.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Boards"
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "500":
          $ref: "#/components/responses/ErrInternalServerError"
      security:
        - cookieAuth: [ ]
  /boards/{id}:
    parameters:
      - schema:
          type: integer
        name: id
        in: path
        required: true
    get:
      tags:
        - Boards
      summary: Returns board by id
      description: Returns board by id
      responses:
        "200":
          description: A JSON object of board.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Board"
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "404":
          $ref: "#/components/responses/ErrBoardNotFound"
        "500":
          $ref: "#/components/responses/ErrInternalServerError"
      security:
        - cookieAuth: [ ]
    patch:
      tags:
        - Boards
      summary: Partial update of board
      description: Partial update of board
      requestBody:
        description: A JSON object containing the data to update board.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartialUpdateBoard"
      responses:
        "200":
          description: A JSON object of updated board.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Board"
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "403":
          $ref: "#/components/responses/ErrForbidden"
        "404":
          $ref: "#/components/responses/ErrBoardNotFound"
      security:
        - cookieAuth: [ ]
    delete:
      tags:
        - Boards
      summary: Delete board
      description: Delete board
      responses:
        "204":
          description: Board deleted successfully.
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "404":
          $ref: "#/components/responses/ErrBoardNotFound"
      security:
        - cookieAuth: [ ]
  /boards/{id}/background:
    parameters:
      - schema:
          type: integer
        name: id
        in: path
        required: true
    put:
      tags:
        - Boards
      summary: Update of board background photo
      description: Update of board background photo
      requestBody:
        description: Multipart form with background photo file
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UpdateBoardBackground"
      responses:
        "200":
          description: A JSON object of updated board.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Board"
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "403":
          $ref: "#/components/responses/ErrForbidden"
        "404":
          $ref: "#/components/responses/ErrBoardNotFound"
      security:
        - cookieAuth: [ ]
...
components:
  schemas:
    Id:
      type: integer
      required: true
      example: 3
    Email:
      type: string
      example: slava@vk.com
    Password:
      type: string
      format: password
      minLength: 8
      maxLength: 100
      example: 12345678
    Username:
      type: string
      minLength: 4
      maxLength: 40
      example: slava51
    Name:
      type: string
      minLength: 1
      maxLength: 60
      example: Slava
...
    MediaSource:
      type: string
      example: https://s3.trello.ru/uploaded_picture.jpg
    Time:
      type: string
      example: "11.07.2023 12:23:45"
...
    SignInReq:
      type: object
      properties:
        username:
          $ref: "#/components/schemas/Username"
        password:
          $ref: "#/components/schemas/Password"
    SignUpReq:
      type: object
      properties:
        name:
          $ref: "#/components/schemas/Name"
        username:
          $ref: "#/components/schemas/Username"
        email:
          $ref: "#/components/schemas/Email"
        password:
          $ref: "#/components/schemas/Password"
    User:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/Id"
        username:
          $ref: "#/components/schemas/Username"
        email:
          $ref: "#/components/schemas/Email"
        name:
          $ref: "#/components/schemas/Name"
        avatar:
          $ref: "#/components/schemas/MediaSource"
        created_at:
          $ref: "#/components/schemas/Time"
        updated_at:
          $ref: "#/components/schemas/Time"
...
    Board:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/Id"
        workspace_id:
          $ref: "#/components/schemas/Id"
        title:
          type: string
          example: "Backend"
        description:
          type: string
          example: "Backend Board"
        background:
          $ref: "#/components/schemas/MediaSource"
        created_at:
          $ref: "#/components/schemas/Time"
        updated_at:
          $ref: "#/components/schemas/Time"
    Boards:
      type: object
      properties:
        boards:
          type: array
          items:
            $ref: "#/components/schemas/Board"
    CreateBoard:
      type: object
      properties:
        title:
          type: string
          required: true
        description:
          type: string
          required: false
      example: {
        "title": "Backend",
        "description": "Backend Board",
      }
    PartialUpdateBoard:
      type: object
      properties:
        workspace_id:
          required: false
          type: integer
          example: 5
        title:
          type: string
          required: false
          example: "Backend"
        description:
          type: string
          required: false
          example: "Backend Board"
    UpdateBoardBackground:
      type: object
      properties:
        background:
          type: string
          format: binary
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
  responses:
    ErrResponseBadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            description: Bad request
            properties:
              error:
                type: string
                example: "{{reason}}"
    ErrUnauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            description: Unauthorized
            properties:
              error:
                type: string
                example: "unauthorized"
    ErrUserNotFound:
      description: User not found
      content:
        application/json:
          schema:
            type: object
            description: User not found
            properties:
              error:
                type: string
                example: "user not found"
...
    ErrBoardNotFound:
      description: Board not found
      content:
        application/json:
          schema:
            type: object
            description: Board not found
            properties:
              error:
                type: string
                example: "board not found"
...
    ErrForbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            description: Access denied
            properties:
              error:
                type: string
                example: "access denied"
    ErrInternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            description: Internal Server Error
            properties:
              error:
                type: string
                example: "{{reason}}"